headers: make sure to forward most target response headers (#19)
This commit forwards all headers from the target server response, but
removes:
- the `set-cookie` header to avoid transfering authentication data via
the proxy (it's not its role to authenticate any clients)
- the `transfer-encoding` header, because it's a [“hop-by-hop”
header](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Transfer-Encoding)
and not supposed to be forwarded. The proxy reads the response (and
should thus consume the `transfer-encoding` header). If one day the
proxy supports streaming responses, we will have to manually add this
header as per the HTTP spec.