CORS: allow requested method & headers values in CORS response (#20)
This is a fix to be able to allow CORS requests when they use specific
headers.

Until now we would limit to a set of headers that we
decided (Content-Type, Authorization, X-Requested-With and the proxy
token header). But this won't work for users needing to send custom
headers in their requests.

Code was inspired by
https://github.com/Rob--W/cors-anywhere/blob/70aaa22b3f9ad30c8566024bf25484fd1ed9bda9/lib/cors-anywhere.js#L59-L66